+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix caching of `__srcIndexArray__`
 - Prevent self-occlusion on quaternary amine
 - Fix outline postprocessing artifacts (black bands) on membrane layers at grazing view angles in Illustrative mode (#1749)
+- Remove fence from `Canvas3D.render` to not interfer with `requestAnimationFrame`
 
 ## [v5.5.0] - 2025-12-22
 - Viewer app